Fixed issue with %p conversion ; printf_mandatory_part done ; Extended road_map with bonus reflexion on realisation

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: gueberso <gueberso@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/14 07:34:07 by gueberso          #+#    #+#             */
-/*   Updated: 2024/11/16 22:09:17 by gueberso         ###   ########.fr       */
+/*   Updated: 2024/11/17 17:25:13 by gueberso         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,9 +17,9 @@ char	*ft_strchr(const char *s, int c);
 
 int		ft_ptrhexa(void *ptr);
 int		ft_putchar(int c);
-int		ft_puthexa_lowercase(unsigned int n);
+int		ft_puthexa_lowercase(unsigned long n);
 int		ft_puthexa_uppercase(unsigned int n);
-int		ft_putnbr_base(unsigned int nbr, char *base);
+int		ft_putnbr_base(unsigned long nbr, char *base);
 int		ft_putnbr(int n);
 int		ft_putstr(char *str);
 int		ft_strlen(const char *s);
